add win allocator.
