+++ runtime/Cargo.toml
@@ -43,6 +43,11 @@ git = "https://github.com/iotaledger/crypto.rs"
 rev = "2dc153079381bf7a87fac54555b26c230a5be593"
 features = [ "random", "chacha", "x25519", "blake2b", "aes" ]
 
+[target.'cfg(windows)'.dependencies]
+sysconf = "0.3"
+winapi = { version = "0.3", features = [ "memoryapi", "sysinfoapi", "winnt" ]}
+getrandom = "0.1"
+
 [build-dependencies]
 bindgen = "0.55.1"
