+++ runtime/src/lib.rs
@@ -18,7 +18,6 @@ extern crate lazy_static;
 #[cfg(feature = "stdalloc")]
 extern crate std;
 
-#[cfg(unix)]
 pub mod mem;
 
 pub mod secret;
@@ -35,7 +34,12 @@ pub enum Error {
         syscall: &'static str,
         errno: libc::c_int,
     },
-    #[cfg(unix)]
+    #[cfg(windows)]
+    OsError {
+        syscall: &'static str,
+        errno: u32,
+    },
+
     MemError(mem::Error),
     ZoneError(zone::Error),
     CryptoError(crypto::Error),
@@ -60,13 +64,22 @@ impl Error {
         }
     }
 
+    #[cfg(target_os = "windows")]
+    pub fn os(syscall: &'static str) -> Self {
+        use winapi::um::errhandlingapi::GetLastError;
+
+        Self::OsError {
+            syscall,
+            errno: { unsafe { GetLastError() } },
+        }
+    }
+
     #[allow(dead_code)]
     fn unreachable(msg: &'static str) -> Self {
         Self::Unreachable(msg)
     }
 }
 
-#[cfg(unix)]
 impl From<mem::Error> for Error {
     fn from(e: mem::Error) -> Self {
         Error::MemError(e)
@@ -80,7 +93,6 @@ impl From<zone::Error> for Error {
     }
 }
 
-#[cfg(unix)]
 impl From<crypto::Error> for Error {
     fn from(e: crypto::Error) -> Self {
         Error::CryptoError(e)
@@ -100,17 +112,44 @@ fn strerror(errno: libc::c_int) -> &'static str {
     }
 }
 
+#[cfg(windows)]
+fn strerror(errno: u32) -> &'static str {
+    use winapi::shared::minwindef::DWORD;
+    use winapi::shared::ntdef::WCHAR;
+    use winapi::um::winbase::{FORMAT_MESSAGE_FROM_SYSTEM, FORMAT_MESSAGE_IGNORE_INSERTS};
+
+    let lang_id = 0x0800 as DWORD;
+
+    let mut buf = [0 as WCHAR; 2048];
+
+    unsafe {
+        let res = winapi::um::winbase::FormatMessageW(
+            FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
+            core::ptr::null_mut(),
+            errno as DWORD,
+            lang_id,
+            buf.as_mut_ptr(),
+            buf.len() as DWORD,
+            core::ptr::null_mut(),
+        );
+
+        assert_ne!(res, 0);
+
+        let len = buf.iter().position(|c| *c == 0).unwrap_or(buf.len());
+
+        core::str::from_utf8_unchecked(core::slice::from_raw_parts(buf.as_ptr() as *const u8, len))
+    }
+}
+
 impl fmt::Debug for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            #[cfg(unix)]
             Self::OsError { syscall, errno } => f
                 .debug_struct("OsError")
                 .field("syscall", syscall)
                 .field("errno", errno)
                 .field("strerror", &strerror(*errno))
                 .finish(),
-            #[cfg(unix)]
             Self::MemError(me) => me.fmt(f),
             Self::ZoneError(ze) => ze.fmt(f),
             Self::CryptoError(ce) => ce.fmt(f),
